refactor : SCRUM-43 DTO 폴더 위치 변경
